Baldur.Server/requirements.txt
alembic==1.0.11
argon2-cffi==19.1.0
cffi==1.12.3
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
parso==0.5.1
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
